Create the initial project structure
This commits the generated code, this should be removed at a later
date, to use it more dynamically or have a way to build it without
much issue.

This project can be ran from a REPL and submit some commands to a
listening Anoma Client.

The protobuf file is currently a WIP and should be refined